requirements.txt
APScheduler==3.6.3
attrdict==2.0.1
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Fabric3==1.14.post1
facebook-wda==0.9.2
Flask==1.1.1
Flask-APScheduler==1.11.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
logzero==1.5.0
MarkupSafe==1.1.1
paramiko==2.7.1
Pillow==7.1.2
py==1.8.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
retry==0.9.2
six==1.14.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
